game architecture refactored, all game logic moved to separate folder
