+++ docs/game.md
@@ -62,7 +62,7 @@ mockGameState.ts - моковый прототип стейта игры (впо
 
 gameEngine/types:
 
-    commonTypes.ts - общие типы игры: точка координат, тип корабля, тип выстрела, а также общий тип отрисовываемого обьекта (DrawableGameObject) у которого есть общий тип состояния с координатами для отрисовки и общий тип параметров с размерами.
+    commonTypes.ts - общие типы игры: точка координат, тип корабля, тип выстрела, а также общий тип отрисовываемого обьекта (BaseObject) у которого есть общий тип состояния с координатами для отрисовки и общий тип параметров с размерами.
 
     gameTypes.ts - игровые типы: общий тип отрисовываемого обьекта, его наследники корабль и выстрел
