+++ packages/client/src/app/pages/startGame/startGame.tsx
@@ -1,10 +1,10 @@
 /* eslint-disable no-plusplus */
 import React, { FC, useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
+import { GlobalGameState } from '@game/store/objectState';
 import Progressbar from '@/app/components/common/progressbar/progressbar';
 import style from './startgame.module.scss';
 import { setGameState } from '@/app/store/slices/gameSlice';
-import { GlobalGameState } from '../game/gameEngine/store/objectState';
 
 const StartGame: FC = () => {
     const [value, setValue] = useState(0);
