+++ packages/client/vite.config.ts
@@ -17,6 +17,7 @@ export default defineConfig({
             '@components': path.resolve(__dirname, './src/components'),
             '@pages': path.resolve(__dirname, './src/pages'),
             '@common': path.resolve(__dirname, './src/components/common'),
+            '@game': path.resolve(__dirname, './src/game'),
         },
     },
     define: {
